requirements.txt
streamlit
PyPDF2
numpy
python-docx
google.generativeai
regex
reportlab
